0017-fixes varios: eliminacion componentes obsoletos y la property onPersonaFund en PersonaFinder
