+++ frontend/src/components/AddJornalComponent.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+/* import React, { useState, useEffect } from 'react'
 import { useNavigate, useParams } from 'react-router-dom';
 import JornalService from '../services/JornalService';
 import ContainerPersonaFinderComponent from './functionalComponents/ContainerPersonaFinderComponent';
@@ -341,3 +341,4 @@ export const AddJornalComponent = () => {
 
 export default AddJornalComponent
 
+ */
