+++ frontend/src/components/Headers/HeaderAdminComponent.js
@@ -49,7 +49,6 @@ const HeaderAdminComponent = () => {
                                 {/*<a className="dropdown-item" href="/jornal">Agregar</a>*/}
                                 <a className="dropdown-item" href="/add-rain">Agregar lluvia</a>
                                 <a className="dropdown-item" href="/buscar-jornal">Buscar</a>
-                                <a className="dropdown-item" href="/buscar-jornal-dos">Buscar Dos</a>
                                 <a className="dropdown-item" href="/confirmar-jornal">Confirmar</a>
                             </div>
                         </li>
@@ -59,8 +58,6 @@ const HeaderAdminComponent = () => {
                             </a>
                             <div className="dropdown-menu" aria-labelledby="navbarDropdownMenuLinkReportes">
                                 <a className="dropdown-item" href="/reporte">Obtener nuevo reporte</a>
-                                <a className="dropdown-item" href="/reporte-dos">Nuevo reporte dos</a>
-                                <a className="dropdown-item" href="#">Historial</a>
                             </div>
                         </li>
