+++ frontend/src/routes/AdminRoutes.js
@@ -7,13 +7,11 @@ import ListObrasComponent from '../components/ListObrasComponent';
 import AddObraComponent from '../components/AddObraComponent';
 import AssignJefeObraComponent from '../components/AssignJefeObraComponent';
 import AssignRolComponent from '../components/AssignRolComponent';
-import ContainerBuscadorJornalComponent from '../components/functionalComponents/ContainerBuscadorJornalComponent';
 import ChangePassword from '../components/functionalComponents/ChangePassword';
-import GenerarReporteComponent from '../components/GenerarReporteComponent'
 import AddRainToObra from '../components/AddRainToObra'
 import AddOrModifyJornalComponent from '../components/AddOrModifyJornalComponent';
 import DeleteJornalComponent from '../components/DeleteJornalComponent';
-import GenerarReportePruebaComponent from '../components/GenerarReportePruebaComponent';
+import GenerarReporteComponent from '../components/GenerarReporteComponent';
 import BuscarJornalComponent from '../components/BuscarJornalComponent';
 import JornalesParaConfirmarAdminComponent from '../components/functionalComponents/JornalesParaConfirmarAdminComponent';
 
@@ -34,11 +32,9 @@ const AdminRoutes = () => {
       <Route exact path='/modify-jornal/:id' element={<AddOrModifyJornalComponent />} />
       <Route exact path='/add-jornal/' element={<AddOrModifyJornalComponent />} />
       <Route exact path='/delete-jornal/:id' element={<DeleteJornalComponent />} />
-      {/* <Route exact path='/buscar-jornal' element={<ContainerBuscadorJornalComponent adminView={true}/>}/> */}
       <Route exact path='/resetpassword/:id' element={<ChangePassword />} />
-      <Route exact path='/reporte' element={<GenerarReporteComponent />} />
       <Route exact path='/add-rain' element={<AddRainToObra/>}/>
-      <Route exact path='/reporte-dos' element={<GenerarReportePruebaComponent/>}/>
+      <Route exact path='/reporte-dos' element={<GenerarReporteComponent/>}/>
       <Route exact path='/buscar-jornal' element={<BuscarJornalComponent showTrabajadores={true} adminView={true}/>}/>
       <Route exact path='/confirmar-jornal' element={<JornalesParaConfirmarAdminComponent/>}/>
     </Routes>
