+++ frontend/src/routes/JefeObraRoutes.js
@@ -2,12 +2,10 @@ import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 import IndexJefeComponent from '../components/Home/IndexJefeComponent';
 import LogoutComponent from '../components/LogoutComponent';
-import ContainerBuscadorJornalComponent from '../components/functionalComponents/ContainerBuscadorJornalComponent';
 import ChangePassword from '../components/functionalComponents/ChangePassword';
 import QrJornalComponent from '../components/QrJornalComponent';
 import AddRainToObra from '../components/AddRainToObra';
 import AddOrModifyJornalComponent from '../components/AddOrModifyJornalComponent';
-import AddJornalComponent from '../components/AddJornalComponent';
 import ListMyJornalesComponent from '../components/ListMyJornalesComponent';
 import BuscarJornalComponent from '../components/BuscarJornalComponent';
 
@@ -18,7 +16,6 @@ const JefeObraRoutes = () => {
       <Route exact path='/home' element={<IndexJefeComponent />} />
       <Route exact path='/logout' element={<LogoutComponent />} />
       <Route exact path='/jornalQr/:id' element={<QrJornalComponent/>}/>
-      {/* <Route exact path='/buscar-jornal' element={<ContainerBuscadorJornalComponent adminView={true}/>}/> */}
       <Route exact path='/resetpassword/:id' element={<ChangePassword />} />
       <Route exact path='/modify-jornal/:id' element={<AddOrModifyJornalComponent />} />
       <Route exact path='/add-rain' element={<AddRainToObra/>}/>
